pcp-poc-app/server/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
azure-core==1.28.0
azure-identity==1.13.0
click==8.1.5
fastapi==0.100.0
h11==0.14.0
httptools==0.6.0
idna==3.4
jinja2==3.1.2
numpy==1.25.1
pandas==2.0.3
pydantic==2.0.3
pydantic-settings==2.0.2
pydantic_core==2.3.0
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn[standard]==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
